+++ packages/component/src/block/accountStatus/desktop/leverage.tsx
@@ -17,7 +17,7 @@ interface LeverageAndMarginRatioProps {
 }
 
 const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
-  const { isConnected } = props;
+  const { isConnected, ...rest } = props;
   const { marginRatio, currentLeverage, mmr } = useMarginRatio();
   const [maxLeverage, { update, config: leverageLevers }] = useLeverage();
 
@@ -26,7 +26,11 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
   const { isRed, isYellow, isGreen } = getMarginRatioColor(marginRatioVal, mmr);
 
   return (
-    <div className={"orderly-flex orderly-justify-between orderly-text-xs"}>
+    <div
+      id="orderly-account-leverage"
+      className={"orderly-flex orderly-justify-between orderly-text-xs"}
+      {...rest}
+    >
       <div
         className={
           "orderly-flex orderly-flex-col orderly-tabular-nums orderly-items-start"
@@ -49,6 +53,7 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
           className="orderly-max-w-[270px]"
         >
           <div
+            id="orderly-account-margin-ratio-title"
             className={
               "orderly-text-base-contrast-54 orderly-text-3xs orderly-cursor-pointer"
             }
@@ -57,6 +62,7 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
           </div>
         </Tooltip>
         <Numeral
+          id="orderly-account-margin-ratio-value"
           className={cn(
             "orderly-text-base-contrast",
             isConnected &&
@@ -72,7 +78,10 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
           {isConnected ? marginRatioVal : "-"}
         </Numeral>
       </div>
-      <div id="orderly-desktop-max-leverage" className={"orderly-flex orderly-flex-col orderly-items-end"}>
+      <div
+        id="orderly-desktop-max-leverage"
+        className={"orderly-flex orderly-flex-col orderly-items-end"}
+      >
         <Tooltip
           content={
             "Your actual Leverage of the whole account / Your max Leverage of the whole account"
@@ -80,6 +89,7 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
           className="orderly-max-w-[270px]"
         >
           <div
+            id="orderly-account-leverage-title"
             className={
               "orderly-text-base-contrast-54 orderly-text-3xs orderly-cursor-pointer"
             }
@@ -88,6 +98,7 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
           </div>
         </Tooltip>
         <div
+          id="orderly-account-leverage-value"
           className={
             "orderly-flex orderly-items-center orderly-gap-1 orderly-text-2xs"
           }
@@ -101,7 +112,10 @@ const LeverageAndMarginRatio = (props: LeverageAndMarginRatioProps) => {
           <span className={"orderly-text-base-contrast-54"}>/</span>
           {isConnected ? (
             <LeverageDialog>
-              <button id="orderly-desktop-leverage-button" className="orderly-flex orderly-items-center orderly-gap-1">
+              <button
+                id="orderly-desktop-leverage-button"
+                className="orderly-flex orderly-items-center orderly-gap-1"
+              >
                 <span>{`${maxLeverage ?? "-"}x`}</span>
                 {typeof maxLeverage !== "undefined" && (
                   // @ts-ignore