+++ packages/component/src/page/trading/mobile/sections/bottombar/index.tsx
@@ -16,6 +16,7 @@ import {
 } from "@orderly.network/hooks";
 import { AccountStatusEnum } from "@orderly.network/types";
 import { isTestnet } from "@orderly.network/utils";
+import TopTips from "@/block/accountStatus/sections/topTips";
 
 interface BottomNavBarProps {}
 
@@ -58,13 +59,7 @@ export const BottomNavBar: React.FC<BottomNavBarProps> = (props) => {
 
   return (
     <>
-      {wsStatus === WsNetworkStatus.Unstable ? (
-        <WsStatus />
-      ) : (
-        errors?.ChainNetworkNotSupport && (
-          <ChainIdSwtich onSetChain={onSetChain} />
-        )
-      )}
+      <TopTips />
 
       {showGetTestUSDC && <GetTestUSDC />}
       <div
@@ -72,7 +67,6 @@ export const BottomNavBar: React.FC<BottomNavBarProps> = (props) => {
         className="orderly-fixed orderly-left-0 orderly-bottom-0 orderly-w-screen orderly-bg-base-800 orderly-p-[14px] orderly-pb-[20px] orderly-border-t orderly-border-base-contrast/10 orderly-z-30 orderly-h-[64px] orderly-flex orderly-justify-between orderly-items-center"
       >
         <AccountStatusBar
-          chains={[]}
           status={state.status}
           address={state.address}
           accountInfo={data}