+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,82 @@
 # examples
 
+## 0.4.12
+
+### Patch Changes
+
+- 11399d1f3: add element id;
+- Updated dependencies [11399d1f3]
+  - @orderly.network/react@1.5.12
+  - @orderly.network/core@1.5.12
+  - @orderly.network/hooks@1.5.12
+  - @orderly.network/web3-onboard@1.5.12
+
+## 0.4.12-alpha.0
+
+### Patch Changes
+
+- add element id;
+- Updated dependencies
+  - @orderly.network/react@1.5.12-alpha.0
+  - @orderly.network/core@1.5.12-alpha.0
+  - @orderly.network/hooks@1.5.12-alpha.0
+  - @orderly.network/web3-onboard@1.5.12-alpha.0
+
+## 0.4.11
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/react@1.5.11
+  - @orderly.network/hooks@1.5.11
+  - @orderly.network/core@1.5.11
+  - @orderly.network/web3-onboard@1.5.11
+
+## 0.4.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/react@1.5.10
+  - @orderly.network/hooks@1.5.10
+  - @orderly.network/core@1.5.10
+  - @orderly.network/web3-onboard@1.5.10
+
+## 0.4.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/react@1.5.10-alpha.0
+  - @orderly.network/hooks@1.5.10-alpha.0
+  - @orderly.network/core@1.5.10-alpha.0
+  - @orderly.network/web3-onboard@1.5.10-alpha.0
+
+## 0.4.9
+
+### Patch Changes
+
+- 187dd07b: fix bugs;
+- Updated dependencies [187dd07b]
+  - @orderly.network/react@1.5.9
+  - @orderly.network/hooks@1.5.9
+  - @orderly.network/core@1.5.9
+  - @orderly.network/web3-onboard@1.5.9
+
+## 0.4.9-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/react@1.5.9-alpha.0
+  - @orderly.network/hooks@1.5.9-alpha.0
+  - @orderly.network/core@1.5.9-alpha.0
+  - @orderly.network/web3-onboard@1.5.9-alpha.0
+
 ## 0.4.8
 
 ### Patch Changes