+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @orderly.network/cli
 
+## 0.3.12
+
+### Patch Changes
+
+- 11399d1f3: add element id;
+- Updated dependencies [11399d1f3]
+  - @orderly.network/core@1.5.12
+
+## 0.3.12-alpha.0
+
+### Patch Changes
+
+- add element id;
+- Updated dependencies
+  - @orderly.network/core@1.5.12-alpha.0
+
+## 0.3.11
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/core@1.5.11
+
+## 0.3.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/core@1.5.10
+
+## 0.3.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/core@1.5.10-alpha.0
+
+## 0.3.9
+
+### Patch Changes
+
+- 187dd07b: fix bugs;
+- Updated dependencies [187dd07b]
+  - @orderly.network/core@1.5.9
+
+## 0.3.9-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/core@1.5.9-alpha.0
+
 ## 0.3.8
 
 ### Patch Changes