+++ packages/referral/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @orderly.network/referral
 
+## 1.5.12
+
+### Patch Changes
+
+- 11399d1f3: add element id;
+- Updated dependencies [11399d1f3]
+  - @orderly.network/react@1.5.12
+  - @orderly.network/hooks@1.5.12
+  - @orderly.network/utils@1.5.12
+
+## 1.5.12-alpha.0
+
+### Patch Changes
+
+- add element id;
+- Updated dependencies
+  - @orderly.network/react@1.5.12-alpha.0
+  - @orderly.network/hooks@1.5.12-alpha.0
+  - @orderly.network/utils@1.5.12-alpha.0
+
+## 1.5.11
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/react@1.5.11
+  - @orderly.network/hooks@1.5.11
+  - @orderly.network/utils@1.5.11
+
+## 1.5.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/react@1.5.10
+  - @orderly.network/hooks@1.5.10
+  - @orderly.network/utils@1.5.10
+
+## 1.5.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/react@1.5.10-alpha.0
+  - @orderly.network/hooks@1.5.10-alpha.0
+  - @orderly.network/utils@1.5.10-alpha.0
+
+## 1.5.9
+
+### Patch Changes
+
+- 187dd07b: fix bugs;
+- Updated dependencies [187dd07b]
+  - @orderly.network/react@1.5.9
+  - @orderly.network/hooks@1.5.9
+  - @orderly.network/utils@1.5.9
+
+## 1.5.9-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/react@1.5.9-alpha.0
+  - @orderly.network/hooks@1.5.9-alpha.0
+  - @orderly.network/utils@1.5.9-alpha.0
+
 ## 1.5.8
 
 ### Patch Changes