+++ packages/walletconnect/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @orderly.network/web3-modal
 
+## 1.5.12
+
+### Patch Changes
+
+- 11399d1f3: add element id;
+
+## 1.5.12-alpha.0
+
+### Patch Changes
+
+- add element id;
+
+## 1.5.11
+
+### Patch Changes
+
+- fix bugs;
+
+## 1.5.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+
+## 1.5.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+
+## 1.5.9
+
+### Patch Changes
+
+- 187dd07b: fix bugs;
+
+## 1.5.9-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+
 ## 1.5.8
 
 ### Patch Changes